requirements.txt
arxiv
langchain
langchain-community
langchain-groq
mcp
nest-asyncio
python-dotenv